Refactor: Refacotr auth middleware and adding type/express on package json and removing the @types folder
