+++ package.json
@@ -10,6 +10,7 @@
     "docker:up": "docker compose up"
   },
   "devDependencies": {
+    "@types/express": "^4.17.21",
     "@types/bcryptjs": "^2.4.6",
     "@types/cors": "^2.8.17",
     "@types/jsonwebtoken": "^9.0.7",
@@ -19,7 +20,6 @@
   },
   "dependencies": {
     "@prisma/client": "^5.21.1",
-    "@types/express": "^4.17.21",
     "@types/swagger-ui-express": "^4.1.6",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
