--- src/@types/express/index.d.ts
@@ -1,6 +1,0 @@
-declare namespace Express {
-    export interface Request {
-        user_id: string;
-        role: string;
-    }
-}
