+++ src/controllers/auth/AuthController.ts
@@ -4,11 +4,13 @@ import { AuthRequest } from '../../interfaces/auth/authRequest';
 
 class AuthController {
     async handle(request: Request, response: Response) {
+
       const authService = new AuthService();
       const { email, username, password, role } = request.body as AuthRequest;
       const auth = await authService.execute({
         email, username, password, role
       });
+
       return response.json(auth);
   }
 }
