+++ src/routes/private/post.private.routes.ts
@@ -6,6 +6,7 @@ import { validateSchema } from "../../middlewares/schema/validateSchema";
 import { createPostSchema } from "../../schemas/postSchemas";
 import { updatePostSchema } from "../../schemas/postSchemas";
 import { isAuthenticated } from "../../middlewares/auth/isAuthenticated";
+import { authMiddleware } from "../../middlewares/auth/Auth.middleware";
 
 /**
  * @module postsPrivateRoutes
@@ -24,7 +25,7 @@ const createPostController = new CreatePostController();
 const updatePostController = new UpdatePostController();
 const deletePostController = new DeletePostController();
 
-postsPrivateRoutes.use(isAuthenticated);
+postsPrivateRoutes.use(authMiddleware);
 
 postsPrivateRoutes.post('/', validateSchema(createPostSchema), createPostController.create);
 postsPrivateRoutes.put('/:id', validateSchema(updatePostSchema), updatePostController.update);
