+++ src/routes/private/tag.private.routes.ts
@@ -1,11 +1,12 @@
 import { Router } from 'express';
 import { isAuthenticated } from '../../middlewares/auth/isAuthenticated'; 
 import { CreateTagController } from '../../controllers/tag/CreateTagController';
+import { authMiddleware } from '../../middlewares/auth/Auth.middleware';
 
 const tagPrivateRoutes = Router();
 
 const createTagController = new CreateTagController();
 
-tagPrivateRoutes.post('/', isAuthenticated, createTagController.create)
+tagPrivateRoutes.post('/', authMiddleware, createTagController.create)
 
 export { tagPrivateRoutes };
